split process manager example
