+++ .gitignore
@@ -25,7 +25,9 @@ examples/logs-lib/basic/basic
 examples/logs-lib/custom/custom
 examples/logs-lib/http/http
 examples/logs-plugin/logs-plugin
-examples/process-manager-plugin/process-manager-plugin
+examples/process-manager-plugin/basic-scenario/basic-scenario
+examples/process-manager-plugin/advanced-scenario/advanced-scenario
+examples/process-manager-plugin/templates/templates
 examples/process-manager-plugin/test-process/test-process
 examples/redis-lib/airport/airport-linux
 examples/redis-lib/simple/simple-linux
