+++ examples/process-manager-plugin/advanced-scenario/Readme.md
@@ -1,0 +1,4 @@
+# Process manager advanced scenario
+
+The example shows how to detach the process from parent application on start, how to attach to running process
+using process ID and work with process status file.
