+++ examples/process-manager-plugin/basic-scenario/Readme.md
@@ -1,0 +1,3 @@
+# Process manager basic scenario
+
+The example shows how to start, restart or stop the process and how to use notification channel to watch process status.
