+++ examples/process-manager-plugin/templates/Readme.md
@@ -1,0 +1,13 @@
+# Process manager template scenario
+
+The example shows how to create or generate a process template, how to use it to run process or how to run
+processes on plugin start up.
+
+Note: in order to run a part of the example which handles templates, process manager config file needs to be provided
+to the example, otherwise it will be skipped.
+
+```
+./process-manager-plugin -process-manager-config=<path-to-file>
+```
+
+All about process manager config file can be found in process manager [readme](../../process/README.md#Templates)
